+++ cmd/repoctl/status.go
@@ -7,7 +7,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/goulash/pacman/aur"
+	"github.com/cassava/repoctl/pacman/aur"
 	"github.com/spf13/cobra"
 )