+++ NEWS.md
@@ -5,6 +5,7 @@ Repoctl Releases
 
 - New: `search` command added.
 - Update: `new config` command now backs up existing configuration files.
+- Bugfix: `version` command does not show entire configuration.
 
 ## Version 0.19 (25 October 2019)
 This release fixes several bugs and adds support for signatures and Zst